app/src/core/security.py
"""
Security utilities for authentication and authorization.

Provides:
- Password hashing and verification
- JWT token creation and validation
- Authentication dependencies
"""
import logging
from datetime import datetime, timedelta
from typing import Optional

from jose import JWTError, jwt
from passlib.context import CryptContext

from .config import settings

logger = logging.getLogger(__name__)

# Password hashing context
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

logger.info("Security module initialized")


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a password against a hash."""
    logger.debug("Verifying password")
    # TODO: Implement password verification
    return False


def get_password_hash(password: str) -> str:
    """Generate password hash."""
    logger.debug("Hashing password")
    # TODO: Implement password hashing
    return ""


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """Create JWT access token."""
    logger.info(f"Creating access token for user: {data.get('sub')}")
    # TODO: Implement JWT token creation
    return ""


def decode_access_token(token: str) -> Optional[dict]:
    """Decode and validate JWT token."""
    logger.debug("Decoding access token")
    # TODO: Implement JWT token decoding
    return None
