app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
asyncffmpeg==1.3.0
pydub==0.25.1
aiogram==3.15.0
aiofiles==24.1.0
python-dotenv==1.0.0
